Refactor: Blacklist customer 따로 분리
